requirements.txt
django==3.1.7
djangorestframework==3.12.3
psycopg2-binary==2.8.4
Pillow==8.1.2
django-filter==2.4.0